+++ app/src/config.ts
@@ -28,6 +28,7 @@ const App: StandaloneAppDescription = {
     "icon.clock": "resources/clock.svg",
     "icon.credit-card": "resources/credit-card.svg",
     "icon.file-text": "resources/file-text.svg",
+    "icon.github": "resources/github.svg",
     "icon.timer": "resources/timer.svg",
     "icon.trending-up": "resources/trending-up.svg",
   },